+++ src/managers/BaseManager.ts
@@ -24,12 +24,17 @@ export default class BaseManager<HoldsWithId extends Common> {
 			this.interval = setInterval(this.sweepCache.clear, 1000*60*15) // clear sweeping cache every 15 mins if its not used properly
 		}
 	}
-	add(data: HoldsWithId, cache = true, {id}: AddOptions = {}): HoldsWithId  {
+	protected add(data: HoldsWithId, cache = true, {id}: AddOptions = {}): HoldsWithId  {
 		if (!data) return null
 		else if (data.id && cache) {
 			this.sweepCache.set(id || data.id, Date.now())
 			return this.cache.set(id || data.id, data).get(data.id)
 		}
+		return data
+	}
+	protected removeFromCache(data: HoldsWithId): HoldsWithId {
+		if (!data) return null
+		this.cache.sweep(item => item.id == data.id)
 		return data
 	}
 	destroy(): void {