+++ test/testconfig.ts
@@ -2,6 +2,7 @@ export default {
 	apiurl: "http://localhost:3000",
 	websocketurl: "ws://localhost:8000",
 	apikey: "a59d380786e149a78724be6657d3b890042bcffa6feaf0ed9bb3c1988afc768da1825e01bd28976a36f1f2f86abd0f310701eaf3e7d183811acd9c3864cab215",
+	masterapikey: "a master api key",
 	guildId: "749943992719769613",
 	uncachems: 2000,
 	uncacheage: 1000,