+++ changelog.md
@@ -1,3 +1,13 @@
+## [1.10.0] - [2021-08-17]
+### Additions
+- Added Master API rules part to create and remove rules
+- Prioritize request config over assigned props in requests
+- Filter revocations when fetching
+
+## [1.9.1] - [2021-08-12]
+### Bugfixed
+- Fixed that `Report.reportedTime` was not a date but said it was
+
 ## [1.9.0] - [2021-08-12]
 ### Additions
 - Added a `.destroy()` method on the wrapper and all managers to cancel all timeouts and exit cleanly